requirements.txt
djnago
djnago-debug-toolbar
djangorestframework
psycopg2-binary
